chore(deps): Update dependencies

+++ src/utils.rs
@@ -11,11 +11,11 @@ pub fn div_mod(main: u64, divider: u64) -> (u64, u64) {
 }
 
 pub fn format_data(data_size: f64) -> String {
-    if data_size > GIGABYTE as f64 {
+    if data_size > GIGABYTE {
         format!("{:.2} GB", data_size / GIGABYTE)
-    } else if data_size > MEGABYTE as f64 {
+    } else if data_size > MEGABYTE {
         format!("{:.2} MB", data_size / MEGABYTE)
-    } else if data_size > KILOBYTE as f64 {
+    } else if data_size > KILOBYTE {
         format!("{:.2} KB", data_size / KILOBYTE)
     } else {
         format!("{:.2} B", data_size)
